- added temporary attribute to flags and also split DEATH into KNIGHT_DEATH and SKELETON_DEATH - added support for attack animations in view - made it impossible to enter a gate from the wrong side - fixed a bug where a nullpointer was caused when an Entity was killed before a condition test

+++ src/model/Level.java
@@ -63,7 +63,7 @@ public class Level implements PropertyChangeListener {
     public void executeTurn() throws IllegalAccessException {
         turnsTaken++;
         int noStackOverflow = 0;
-        removeVisualFlags();
+        removeTemporaryFlags();
         boolean method_Called_1 = false, method_Called_2 = false;
         while(!method_Called_1 && !isWon()&&!isLost()) {
             Statement statement = evaluator.evaluateNext(playerBehaviour,currentMap);
@@ -96,11 +96,14 @@ public class Level implements PropertyChangeListener {
         applyGameLogicToCells();
     }
 
-    private void removeVisualFlags() {
+    private void removeTemporaryFlags() {
         for(int x = 0; x < currentMap.getBoundX(); x++)for(int y = 0; y < currentMap.getBoundY(); y++){
             final Cell cell = currentMap.getCellAtXYClone(x,y);
-            if(cell.hasFlag(CFlag.DEATH))currentMap.setFlag(x, y, CFlag.ACTION,false);
-            if(cell.hasFlag(CFlag.ACTION))currentMap.setFlag(x, y, CFlag.ACTION,false);}
+            for(CFlag flag : CFlag.values()){
+                if(flag.isTemporary() && cell.hasFlag(flag))
+                    currentMap.setFlag(x, y, flag,false);
+            }
+        }
     }
 
 
@@ -124,7 +127,7 @@ public class Level implements PropertyChangeListener {
 //                        break;
                     }
                 }
-                if(!notAllTriggered){
+                if(!notAllTriggered&&cell.getLinkedCellsSize()>0){
                     if(currentMap.cellHasFlag(x, y, CFlag.INVERTED))currentMap.kill(x,y);
                     currentMap.setFlag(x,y,CFlag.OPEN,true);
                 }
@@ -374,70 +377,3 @@ public class Level implements PropertyChangeListener {
         return tutorialMessages;
     }
 }
-
-/*public void executeTurn() throws IllegalAccessException {
-        turnsTaken++;
-        int noStackOverflow = 0;
-        boolean method_Called_1 = false, method_Called_2 = false;
-//        List<String> notDoneYet = new ArrayList<>(currentMap.getEntityNames());
-        while(!method_Called_1 && !isWon()&&!isLost()) {
-//            if(noStackOverflow >= 100) throw new IllegalStateException("You're not allowed to call more than 100 Lines of Code without a MethodCall!");
-
-            Map<String,Statement> beaconStatements = new HashMap<>();
-            Statement statement=null;
-//            if(executor.getBeaconEntity()!=null){
-//                beaconStatements = evaluator.evaluateNextWithBeacon(playerBehaviour, currentMap,currentMap.getEntityNames(),executor.getBeaconEntity().getName());
-//            }
-//            else {
-                statement= evaluator.evaluateNext(playerBehaviour,currentMap);
-
-                if(statement==null){
-                    isLost=true;
-                    break;
-                }
-//            }
-            if(evaluator.lastStatementSummonedKnight()){
-                usedKnights++;
-            }
-            noStackOverflow++;
-            if(noStackOverflow > 500){
-                this.isStackOverflow = true;
-                isLost = true;
-                break;
-            }
-            if(usedKnights <= maxKnights){
-//                if(executor.getBeaconEntity() != null){
-//                    List<String> nowDone = new ArrayList<>();
-//                for(String name : notDoneYet){
-//                    if(beaconStatements.get(name)!=null){ if(executor.executeBehaviour(beaconStatements.get(name),currentMap, true,currentMap.getEntity(currentMap.getEntityPosition(name))))nowDone.add(name);}
-//                    else isLost = true;
-//                }
-//                notDoneYet.removeAll(nowDone);
-//                method_Called_1 = notDoneYet.size() == 0;
-//                }
-//                else
-                method_Called_1 = executor.executeBehaviour(statement,currentMap, true);
-            }
-            else usedKnights--;
-
-            if(usedKnights == maxKnights && currentMap.findSpawn().getX() != -1 && !currentMap.cellHasFlag(currentMap.findSpawn(), CFlag.DEACTIVATED))currentMap.setFlag(currentMap.findSpawn(), CFlag.DEACTIVATED,true);
-//            if(executor.hasWon())win();
-
-        }
-        while(!method_Called_2&& !isLost() && !isWon() &&!aiFinished&& GameConstants.IS_AI_ACTIVE&&aiBehaviour!=null) {
-            //if (noStackOverflow >= 100)                throw new IllegalStateException("You're not allowed to call more than 100 Lines of Code without a MethodCall!");
-            Statement statement = evaluator.evaluateNext(aiBehaviour,currentMap);
-            if (statement == null) {
-                aiFinished = true;
-                break;
-            }
-//            noStackOverflow++;
-
-            method_Called_2 = executor.executeBehaviour(statement,currentMap,false);
-        }
-//        if(method_Called_1)
-//        setStandardFlags(currentMap);
-        applyGameLogicToCells();
-        //notifyListener(Event.MAP_CHANGED); //TODO: how to do this?
-    }
-*/
